Update invitation data class
