+++ app/src/main/java/com/android3/siegertpclient/data/invitation/Invitation.kt
@@ -1,9 +1,11 @@
-package com.android3.siegertpclient.data.invitation
+package com.android3.siegertpclient.data.invitation.invitationsource
 
 data class Invitation(
     val invitationId: String,
     val participantForm: String,
     val recipientId: String,
     val senderId: String,
-    val tournamentId: String
+    val senderUsername: String,
+    val tournamentId: String,
+    val tournamentName: String
 )
