+++ app/src/main/java/com/android3/siegertpclient/data/team/teamsource/TeamRepo.kt
@@ -1,7 +1,7 @@
 package com.android3.siegertpclient.data.team.teamsource
 
 import android.content.Context
-import com.android3.siegertpclient.data.invitation.Invitation
+import com.android3.siegertpclient.data.invitation.invitationsource.Invitation
 import com.android3.siegertpclient.data.payload.ApiResponse
 import com.android3.siegertpclient.data.team.Team
 import com.android3.siegertpclient.data.team.teamsource.teamRemote.TeamRemoteDataSource
