+++ app/src/main/java/com/android3/siegertpclient/data/team/teamsource/teamRemote/TeamService.kt
@@ -1,6 +1,6 @@
 package com.android3.siegertpclient.data.team.teamsource.teamRemote
 
-import com.android3.siegertpclient.data.invitation.Invitation
+import com.android3.siegertpclient.data.invitation.invitationsource.Invitation
 import com.android3.siegertpclient.data.payload.ApiResponse
 import com.android3.siegertpclient.data.team.Team
 import com.android3.siegertpclient.data.tournament.Tournament
