--- app/src/main/java/com/android3/siegertpclient/data/tournament/tournamentsource/ITournamentDataSource.kt
@@ -1,4 +1,0 @@
-package com.android3.siegertpclient.data.tournament.tournamentsource
-
-interface ITournamentDataSource {
-}
