+++ app/src/main/java/com/android3/siegertpclient/data/tournament/tournamentsource/tournamentRemote/TournamentRemoteDataSource.kt
@@ -1,94 +1,79 @@
-//package com.android3.siegertpclient.data.tournament.tournamentsource.tournamentRemote
-//
-//import com.android3.siegertpclient.data.game.Game
-//import com.android3.siegertpclient.data.tournament.Tournament
-//import com.android3.siegertpclient.data.tournament.tournamentsource.ITournamentDataSource
-//import com.android3.siegertpclient.data.user.User
-//import java.util.*
-//
-//class TournamentRemoteDataSource (private val tournamentService: TournamentService) : ITournamentDataSource {
-//
-//    private fun convertTRespToTournament() : Tournament? {
-//        return null
-//    }
-//
-//    fun createNewTournament(tournamentForm : String, tournamentSize : String, tournamentName: String, ownUserId: String) : Tournament? {
-//        return null
-//    }
-//
-//    fun getTournamentById(tourneyId : String, ownUserId: String) : Tournament? {
-//        val userCall = tournamentService.getTournamentById(tourneyId, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return null
-//    }
-//
-//    fun getTournamentByName(tournamentName : String, ownUserId: String) : Tournament? {
-//        val userCall = tournamentService.getTournamentByName(tournamentName, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return null
-//    }
-//
-//    fun getTournamentParticipants(tournamentName : String, ownUserId: String) : Array<User>? {
-//        val userCall = tournamentService.getTournamentParticipants(tournamentName, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return response.body()
-//    }
-//
-//    fun updateTournamentDetailById(tournamentName : String, adminId : String, typeOfGame : String, location : String,
-//                                   registrationDeadline : Date, startTime : Date, endTime : Date,
-//                                   ownUserId: String) : Tournament? {
-//        return null
-//    }
-//
-//    fun deleteTournament(tournamentName : String, ownUserId: String) : Boolean? {
-//        val userCall = tournamentService.deleteTournament(tournamentName, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return response.body()
-//    }
-//
-//    fun handleParticipation(tournamentName : String, participate : Map<String, Boolean>, ownUserId: String) : Boolean? {
-//        val userCall = tournamentService.handleParticipation(tournamentName, participate, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return response.body()
-//    }
-//
-//    fun getTournamentsGames(tournamentName : String, ownUserId: String) : Array<Game>? {
-//        val userCall = tournamentService.getTournamentsGames(tournamentName, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return response.body()
-//    }
-//
-//    fun createGames(tournamentName : String, ownUserId: String) : Array<Game>? {
-//        val userCall = tournamentService.createGames(tournamentName, ownUserId)
-//        val response = userCall.execute()
-//        if (!response.isSuccessful) {
-//
-//            //TOdo implement error code
-//        }
-//        return response.body()
-//    }
-//}
+package com.android3.siegertpclient.data.tournament.tournamentsource.tournamentRemote
+
+import com.android3.siegertpclient.data.payload.ApiResponse
+import com.android3.siegertpclient.data.team.Team
+import com.android3.siegertpclient.data.tournament.Game
+import com.android3.siegertpclient.data.tournament.Tournament
+import com.android3.siegertpclient.data.tournament.TournamentDetail
+import com.android3.siegertpclient.data.user.User
+import com.android3.siegertpclient.utils.RestClient
+import retrofit2.Response
+
+class TournamentRemoteDataSource {
+
+    suspend fun createNewTournament(
+        type : String,
+        participantSize : Int,
+        name : String,
+        tournamentDetail : TournamentDetail,
+        token : String) : Response<Tournament> {
+        val tournament = hashMapOf<String, Any>()
+        tournament["type"] = type
+        tournament["maxParticipantNumber"] = participantSize
+        tournament["tournamentName"] = name
+        tournament["tournamentDetail"] = tournamentDetail
+        return RestClient.tournamentService.createNewTournament(tournament, token)
+    }
+
+    suspend fun getTournamentById(tourneyId : String, token: String) : Response<Tournament> {
+        return RestClient.tournamentService.getTournamentById(tourneyId, token)
+    }
+
+    suspend fun getTournamentByName(tournamentName : String, token: String) : Response<Tournament> {
+        return RestClient.tournamentService.getTournamentByName(tournamentName, token)
+    }
+
+    suspend fun getTournamentParticipantsUser(tournamentName : String, token: String) : Response<List<User>> {
+        return RestClient.tournamentService.getTournamentParticipantsUser(tournamentName, token)
+    }
+
+    suspend fun getTournamentParticipantsTeam(tournamentName : String, token: String) : Response<List<Team>> {
+        return RestClient.tournamentService.getTournamentParticipantsTeam(tournamentName, token)
+    }
+
+    suspend fun getTournamentGames(tournamentName : String, token: String) : Response<List<Game>> {
+        return RestClient.tournamentService.getTournamentGames(tournamentName, token)
+    }
+
+    suspend fun getGameById(tournamentName : String, gameId : String,  token: String) : Response<Game> {
+        return RestClient.tournamentService.getGameById(tournamentName, gameId, token)
+    }
+
+    suspend fun handleParticipation(tournamentName: String, participation : String, token: String) : Response<ApiResponse> {
+        val participationBody = hashMapOf<String, String>()
+        participationBody["participation"] = participation
+        return RestClient.tournamentService.handleParticipation(tournamentName, participationBody, token)
+    }
+
+    suspend fun createGames(tournamentName : String, token: String) : Response<List<Game>> {
+        return RestClient.tournamentService.createGames(tournamentName, token)
+    }
+
+    suspend fun updateGameById(tournamentName : String, game : Game, token: String) : Response<Game> {
+        return RestClient.tournamentService.updateGameById(tournamentName, game.gameId, game, token)
+    }
+
+    suspend fun deleteTournament(tournamentName : String, token: String) : Response<ApiResponse> {
+        return RestClient.tournamentService.deleteTournament(tournamentName, token)
+    }
+
+    suspend fun deleteGameById(tournamentName : String, gameId : String, token: String) : Response<ApiResponse> {
+        return RestClient.tournamentService.deleteGameById(tournamentName, gameId, token)
+    }
+
+    suspend fun updateTournamentDetail(tournamentName: String, tournamentDetail: TournamentDetail, token: String) : Response<Tournament> {
+        return RestClient.tournamentService.updateTournamentDetail(tournamentName, tournamentDetail, token)
+    }
+
+
+}
