+++ app/src/main/java/com/android3/siegertpclient/ui/createtournament/CreateTournamentPresenter.kt
@@ -1,14 +1,14 @@
 package com.android3.siegertpclient.ui.createtournament
 
 import android.content.Context
-import com.android3.siegertpclient.data.tournament.tournamentsource.TournamentRepo2
+import com.android3.siegertpclient.data.tournament.tournamentsource.TournamentRepo
 import com.android3.siegertpclient.ui.base.BasePresenter
 import com.android3.siegertpclient.utils.OnlineChecker
 
 class CreateTournamentPresenter(private val context: Context) : BasePresenter<CreateTournamentContract.ICreateTournamentView>(), CreateTournamentContract.ICreateTournamentPresenter{
     private var onlineChecker = OnlineChecker(context)
 
-    private var tournamentRepo = TournamentRepo2(context)
+    private var tournamentRepo = TournamentRepo(context)
 
     override fun onBackBtnClicked() {
         view?.navigateToHomepageActivity()
