+++ app/src/main/java/com/android3/siegertpclient/utils/RestClient.kt
@@ -2,7 +2,7 @@ package com.android3.siegertpclient.utils
 
 import com.android3.siegertpclient.data.invitation.invitationsource.invitationRemote.InvitationService
 import com.android3.siegertpclient.data.team.teamsource.teamRemote.TeamService
-import com.android3.siegertpclient.data.tournament.tournamentsource.tournamentRemote.TournamentService2
+import com.android3.siegertpclient.data.tournament.tournamentsource.tournamentRemote.TournamentService
 import com.android3.siegertpclient.data.user.usersource.userRemote.UserService
 import com.android3.siegertpclient.utils.Constants.Companion.BASE_URL
 import okhttp3.OkHttpClient
@@ -39,7 +39,7 @@ object RestClient {
         retrofit.create(TeamService::class.java)
     }
 
-    val tournamentService: TournamentService2 by lazy {
-        retrofit.create(TournamentService2::class.java)
+    val tournamentService: TournamentService by lazy {
+        retrofit.create(TournamentService::class.java)
     }
 }
